Factor out renderGroupRulers.
